refactor: move reusable TestSupport to a utils
